refactor(web): extracts treeshaker into tslib helper-project
