+++ common/web/tslib/package.json
@@ -2,6 +2,16 @@
   "name": "@keymanapp/tslib",
   "description": "An ES5 + esbuild-compatible wrapper for the 'tslib' library",
   "main": "./build/index.js",
+  "exports": {
+    ".": {
+      "types": "./build/index.d.ts",
+      "import": "./build/index.js"
+    },
+    "./esbuild-tools": {
+      "types": "./build/esbuild-tools.d.ts",
+      "import": "./build/esbuild-tools.js"
+    }
+  },
   "scripts": {
     "build": "gosh ./build.sh",
     "clean": "tsc -b --clean",
@@ -10,6 +20,9 @@
   "dependencies": {
     "tslib": "^2.5.2",
     "typescript": "^4.9.5"
+  },
+  "devDependencies": {
+    "esbuild": "^0.15.16"
   },
   "type": "module"
 }
