+++ common/web/tslib/tsconfig.json
@@ -4,18 +4,19 @@
     "allowJs": true,
     "module": "es6",
     "moduleResolution": "Node",
+    "allowSyntheticDefaultImports": true,
     "inlineSources": true,
     "sourceMap": true,
     "declaration": true,
     "target": "es5",
     "tsBuildInfoFile": "./build/tsconfig.tsbuildinfo",
     "types": ["node"],
-    "lib": ["es6"],
+    "lib": ["es6", "DOM"],  // The latter is b/c esbuild expects the type.
     "baseUrl": "./src",
     "outDir": "./build/",
     "rootDir": "./src"
   },
   "include": [
-    "./src/index.ts"
+    "./src/*.ts"
   ]
 }
